SuccessStories: Moved models under it's own package
