+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/JobPostFragment.java
@@ -11,6 +11,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import com.example.jobs4smcyouth.Models.JobListing;
 import com.firebase.client.DataSnapshot;
 import com.firebase.client.Firebase;
 import com.firebase.client.FirebaseError;
