+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/SuccessStoryAdapter.java
@@ -7,6 +7,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import com.example.jobs4smcyouth.Models.SuccessStory;
 import com.example.jobs4smcyouth.Utilities.ExpandableTextView;
 
 import java.util.ArrayList;
