+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/SuccessStoryFragment.java
@@ -3,15 +3,15 @@ package com.example.jobs4smcyouth;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.jobs4smcyouth.Models.SuccessStory;
+
 import java.util.ArrayList;
-import java.util.Arrays;
 
 
 /**
